♻️ Refactor: Layout 변경 리펙토링
#25
